move sugared_value headers
